+++ srvcheck/chains/astar.py
@@ -1,7 +1,7 @@
-from srvcheck.tasks.task import hours, minutes
+from srvcheck.tasks.task import minutes
 from ..tasks import Task
 from .substrate import Substrate
-from .substrate import TaskRelayChainStuck, TaskBlockProductionReportCharts
+from .substrate import TaskRelayChainStuck, TaskBlockProductionReportParachain, TaskBlockProductionReportCharts
 from ..notification import Emoji
 
 class TaskBlockProductionCheck(Task):
@@ -23,51 +23,6 @@ class TaskBlockProductionCheck(Task):
 				elif self.prev == block:
 					return self.notify(f'no block produced in the latest 30 minutes! Last block produced was {self.prev} {Emoji.BlockMiss}')
 				self.prev = block
-		return False
-
-class TaskBlockProductionReportParachain(Task):
-	def __init__(self, services, checkEvery=minutes(10), notifyEvery=hours(1)):
-		super().__init__('TaskBlockProductionReportParachain',
-                    services, checkEvery, notifyEvery)
-		self.prev = None
-		self.prevBlock = None
-		self.lastBlockChecked = None
-		self.totalBlockChecked = 0
-		self.oc = 0
-		self.prevSessionLastBlock = 0
-
-	@staticmethod
-	def isPluggable(services):
-		return services.chain.isParachain()
-
-	def run(self):
-		session = self.s.chain.getSession()
-        
-		if self.prev is None:
-			self.prev = session
-
-		if self.s.chain.isCollating():
-			currentBlock = self.s.chain.getHeight()
-			blocksToCheck = [b for b in self.s.chain.getExpectedBlocks() if b <= currentBlock and (self.lastBlockChecked is None or b > self.lastBlockChecked)]
-			for b in blocksToCheck:
-				a = self.s.chain.getBlockAuthor(b)
-				collator = self.s.conf.getOrDefault('chain.validatorAddress')
-				if a.lower() == collator.lower():
-					self.oc += 1
-				self.lastBlockChecked = b
-				self.totalBlockChecked += 1
-
-		if self.prev != session:
-			self.s.persistent.timedAdd(self.s.conf.getOrDefault('chain.name') + '_sessionBlocksProduced', self.prev)
-			self.s.persistent.timedAdd(self.s.conf.getOrDefault('chain.name') + '_blocksChecked', self.totalBlockChecked)
-			self.prev = session
-			perc = 0
-			if self.totalBlockChecked > 0:
-				perc = self.oc / self.totalBlockChecked * 100
-				self.totalBlockChecked = 0
-			self.s.persistent.timedAdd(self.s.conf.getOrDefault('chain.name') + '_blocksProduced', self.oc)
-			self.s.persistent.timedAdd(self.s.conf.getOrDefault('chain.name') + '_blocksPercentageProduced', perc)
-			self.oc = 0
 		return False
 
 class Astar(Substrate):
@@ -106,3 +61,6 @@ class Astar(Substrate):
 			result = si.query(module='CollatorSelection', storage_function='LastAuthoredBlock', params=[collator])
 			return result.value
 		return 0
+
+	def getStartingRoundBlock(self):
+		return -1
